added streaming for the contributor data from github
