+++ src/routes/+page.svelte
@@ -7,6 +7,9 @@
 	} from "$layout";
 	import { defaultI18nValues, Metadata } from "$lib";
 	import { _ } from "svelte-i18n";
+
+	export let data;
+	$: ({ contributors } = data);
 </script>
 
 <Metadata title={$_("metadata.home", defaultI18nValues)} />
@@ -17,4 +20,4 @@
 
 <ThemesSection />
 
-<CommunitySection />
+<CommunitySection {contributors} />
